+++ lib/fifteen_five/report.rb
@@ -2,5 +2,6 @@ module FifteenFive
   class Report < ApiResource
     # Associations
     belongs_to :user
+    belongs_to :reviewer, class_name: "User"
   end
 end